externalize params

+++ autozoom.py
@@ -90,9 +90,9 @@ if __name__ == '__main__':
 	flvZoom = 1.25
 
 	if arguments_strCenterU != '':
-		flvCenterU = int(arguments_strCenterU)
+		flvCenterU = float(arguments_strCenterU)
 	if arguments_strCenterV != '':
-		flvCenterV = int(arguments_strCenterV)
+		flvCenterV = float(arguments_strCenterV)
 	if arguments_shift != '':
 		flvZoom = float(arguments_shift)
 	if arguments_zoom != '':
